refactor(settings): move schema (keys, defaults, Defaults, SecretKeys) to keys.go (PRD #1021 M1)
Pure motion, same package. Cut the settings schema verbatim from settings.go
into a new keys.go: the KEY const block, the DEFAULTS const block, the Defaults
map, the SecretKeys set, and IsSecret/Known. No exported-API change, no behavior
change, no test edit.

The bounds consts (healthSecondsMin/Max, maxHostedWorkerQuota,
maxJudgeDailyBudget, maxMrReworkCap, maxLabelLen) stay in settings.go for later
milestones. keys.go picks up only the theme import the moved Defaults entry needs;
the three //nolint:gosec G101 directives on the Slack/agent-source key consts move
with the block and still suppress.

Byte-identity confirmed: `go doc -all ./internal/settings` is unchanged, and the
moved content diffs verbatim against the base under --color-moved. Tests
unchanged and green.
